Lanzamiento versión 1.1.0

+++ src/main/java/es/uclm/delivery/persistencia/EntidadDAO.java
@@ -13,13 +13,18 @@ public abstract class EntidadDAO<E> {
     @PersistenceContext
     protected EntityManager entityManager;
 
+    private static final String ENTITY_CANNOT_BE_NULL = "Entity cannot be null";
+
     private Class<E> entityClass;
 
-    public EntidadDAO(Class<E> entityClass) {
+    protected EntidadDAO(Class<E> entityClass) {
         this.entityClass = entityClass;
     }
 
     public int insert(E entity) {
+        if (entity == null) {
+            throw new IllegalArgumentException(ENTITY_CANNOT_BE_NULL);
+        }
         try {
             entityManager.persist(entity);
             return 1; // Success
@@ -28,8 +33,12 @@ public abstract class EntidadDAO<E> {
             return 0; // Failure
         }
     }
+    
 
     public int update(E entity) {
+        if (entity == null) {
+            throw new IllegalArgumentException(ENTITY_CANNOT_BE_NULL);
+        }
         try {
             entityManager.merge(entity);
             return 1; // Success
@@ -40,6 +49,9 @@ public abstract class EntidadDAO<E> {
     }
 
     public int delete(E entity) {
+        if (entity == null) {
+            throw new IllegalArgumentException(ENTITY_CANNOT_BE_NULL);
+        }
         try {
             entityManager.remove(entityManager.contains(entity) ? entity : entityManager.merge(entity));
             return 1; // Success
